requirements.txt
fastapi==0.85.2
Pillow==9.3.0
python-multipart==0.0.5
uvicorn==0.19.0